backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
